// src/platform/plugins/shared/discover/public/context_awareness/types.d.ts
export interface DefaultAppStateColumn {
    /**
     * The field name of the column
     */
    name: string;
    /**
     * The width of the column in pixels -- leave undefined for auto width
     */
    width?: number;
}
/**
 * Parameters passed to the default app state extension
 */
export interface DefaultAppStateExtensionParams {
    /**
     * The current data view
     */
    dataView: DataView;
}
/**
 * Supports customizing the default Discover application state
 */
export interface DefaultAppStateExtension {
    /**
     * The columns to display in the data grid
     */
    columns?: DefaultAppStateColumn[];
    /**
     * The height of each row in the data grid:
     * * -1: auto height mode
     * * 0: single line mode
     * * 1-20: number of lines to display
     */
    rowHeight?: number;
    /**
     * The field to apply for the histogram breakdown
     */
    breakdownField?: string;
    /**
     * The state for chart visibility toggle
     */
    hideChart?: boolean;
    /**
     * The state for data table visibility toggle
     */
    hideTable?: boolean;
}